requirements.txt
certifi==2017.4.17
chardet==3.0.4
deuces==0.2
dj-database-url==0.4.1
djangorestframework==3.6.3
Django==1.11.2
django-markdown-deux==1.0.5
gunicorn==19.6.0
idna==2.5
markdown2==2.3.4
psycopg2==2.7.1
pytz==2017.2
requests==2.18.1
urllib3==1.21.1
whitenoise==2.0.6
